+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+#### 0.1.2
+
+- Add optional `--output-name` for pdf output (default to #{repository_name}.pdf)
+
 #### 0.1.1
 
 - Fix style with the help of rubocop
+- Add full link to Github for sample screenshot to make it work with rubydoc.info
 - Misc code cleanup
 
 #### 0.1.0